% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{criterion}
\alias{criterion}
\title{Computing the GMM criterion.}
\usage{
criterion(par, data, weight, equal)
}
\arguments{
\item{par}{vector of parameter values}

\item{data}{dataframe that contains Y, D, Z, and V with indicators}

\item{weight}{weighting matrix for GMM estimation}

\item{equal}{logical whether the misclassification probabilities do not depend on instrument}
}
\description{
\code{criterion} is an auxiliary function to implement the GMM inference on your own.
It returns a value of the GMM criterion.
}
