% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{computing standard errors and confidence intervals}
\usage{
inference(par, data, weight, equal)
}
\arguments{
\item{par}{vector of parameter values}

\item{data}{dataframe that contains Y, D, Z, and V with indicators}

\item{weight}{weighting matrix for pilot GMM estimation}

\item{equal}{logical whether misclassification probabilities do not depend on Z}
}
\value{
list that contains the following elements.
\item{estimate}{parameter estimates}
\item{se}{standard errors}
\item{ci}{95 percent confidence intervals}
}
\description{
\code{inference} is an auxiliary function to implement the GMM inference on your own.
It returns estimates, standard errors, and confidence intervals
for the parameters in the GMM estimation.
See the package vignette via `vignette("LATEerror")` for more details.
}
