% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{Jtest}
\alias{Jtest}
\title{conducting over-identification test (J test)}
\usage{
Jtest(par, data, equal)
}
\arguments{
\item{par}{GMM estimates}

\item{data}{dataframe containing the outcome, treatment, instrument, exogenous variable with indicators}

\item{equal}{logical whether the misclassification probabilities do not depend on instrument}
}
\value{
vector that contains a test statistic value and p-value
}
\description{
conducting over-identification test (J test)
}
