% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATEerror.R
\name{LATEerror}
\alias{LATEerror}
\title{GMM inference to estimate the LATE for a mismeasured binary treatment}
\usage{
LATEerror(Y, D, Z, V, weight = NULL, optimal = FALSE, equal = TRUE,
  lower = NULL, upper = NULL, control = NULL)
}
\arguments{
\item{Y}{vector of the outcome}

\item{D}{vector of the treatment}

\item{Z}{vector of the instrument}

\item{V}{vector of the exogenous variable}

\item{weight}{weighting matrix for the pilot GMM estimation}

\item{optimal}{logical whether implementing the optimal GMM estimation}

\item{equal}{logical whether the misclassification probabilities do not depend on the instrument}

\item{lower}{vector of lower bounds of the parameters}

\item{upper}{vector of upper bounds of the parameters}

\item{control}{list of the control parameters for \code{\link[DEoptim]{DEoptim}}}
}
\value{
list that contains the two lists: (i) \code{pilot}, (ii) \code{optimal}.\cr
\code{pilot} contains the following elements.
\item{estimate}{vector of parameter estimates}
\item{se}{vector of standard errors}
\item{ci}{matrix that contains 95 percent confidence intervals}
\item{optim}{list that contains the results of the optimization via \code{DEoptim}}
\item{weight}{weighting matrix for the pilot GMM estimation}
\code{optimal} contains the following elements.
\item{estimate}{vector of parameter estimates}
\item{se}{vector of standard errors}
\item{ci}{matrix that contains 95 percent confidence intervals}
\item{overidentification}{vector that contains a value of the test statistic and p-value for the over-identification test}
\item{weight}{weighting matrix for the optimal GMM estimation}
\item{optim}{list that contains the results of the optimization via \code{DEoptim}}
}
\description{
\code{LATEerror} implements the GMM inference to correct a measurement error
problem due to a mismeasured binary treatment in the LATE inference
developed in Yanagi (2017).
See the package vignette via \code{vignette("LATEerror")} for more details.
}
\seealso{
\code{\link[DEoptim]{DEoptim}} for the elements in \code{control}
}
