% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{number_name_par}
\alias{number_name_par}
\title{returning the names and the numbers of the parameters}
\usage{
number_name_par(equal, K)
}
\arguments{
\item{equal}{logical whether misclassification probabilities do not depend on Z}

\item{K}{the number of the elements that V takes}
}
\description{
returning the names and the numbers of the parameters
}
