% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{var_moments}
\alias{var_moments}
\title{Estimating variance-covariance matrix of the moments}
\usage{
var_moments(par, data, equal)
}
\arguments{
\item{par}{GMM estimates}

\item{data}{dataframe containing the outcome, treatment, instrument, exogenous variable}

\item{equal}{logical whether the misclassification probabilities do not depend on instrument}
}
\description{
Estimating variance-covariance matrix of the moments
}
