% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{criterion}
\alias{criterion}
\title{Computing the value of the GMM criterion.}
\usage{
criterion(par, data, weight, equal)
}
\arguments{
\item{par}{A vector of parameters for the GMM estimation.}

\item{data}{A data.frame containing the outcome Y, treatment D, instrument Z, exogenous variable V.}

\item{weight}{A matrix for the weighting matrix in the GMM estimation.}

\item{equal}{Logical. Whether imposing restrictions that the misclassification probabilities do not depend on Z?}
}
\description{
Computing the value of the GMM criterion.
}
