% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LATEerror.R, R/package_document.R
\docType{package}
\name{LATEerror}
\alias{LATEerror}
\alias{LATEerror}
\alias{LATEerror-package}
\title{Implementing the GMM inference to estimate the LATE for misclassified treatment.}
\usage{
LATEerror(Y, D, Z, V, weight = NULL, optimal = TRUE, equal = TRUE,
  lower = NULL, upper = NULL, controlDE = NULL, controlBFGS = NULL)
}
\arguments{
\item{Y}{A vector of outcome variable.}

\item{D}{A vector of treatment variable.}

\item{Z}{A vector of instrument.}

\item{V}{A vector of exogenous variable.}

\item{weight}{A matrix for the weighting matrix for pilot GMM estimation.}

\item{optimal}{Logical. Whether implementing the optimal GMM estimation?}

\item{equal}{Logical. Whether imposing restrictions that the misclassification probabilities do not depend on Z?}

\item{lower}{A vector for lower bounds of parameters.}

\item{upper}{A vector for upper bounds of parameters.}

\item{controlDE}{A list for control parameters for \code{\link[DEoptim]{DEoptim}}.}

\item{controlBFGS}{A list for control parameters for \code{\link[stats]{optim}} with L-BFGS-B method.}
}
\value{
The output is a list that contains the following two list.
  (1) A list for pilot estimation results by the pilot GMM estimation based on weight.
  (2) A list for optimal estimation results by the optimal GMM estimation based on the optimal weighting matrix.
}
\description{
Implementing the GMM inference to estimate the LATE for misclassified treatment.

The LATEerror package provides \code{\link{LATEerror}} function.
}
\seealso{
\code{\link[DEoptim]{DEoptim}} for controlDE and \code{\link[stats]{optim}} for controlBFGS.
}
