% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Computing the values of the moments for the GMM estimation.}
\usage{
moments(par, data, equal)
}
\arguments{
\item{par}{A vector of parameters for the GMM estimation.}

\item{data}{A data.frame containing the outcome Y, treatment D, instrument Z, exogenous variable V.}

\item{equal}{Logical. Whether imposing restrictions that the misclassification probabilities do not depend on Z?}
}
\description{
Computing the values of the moments for the GMM estimation.
}
